Update README.md
JP Software CMDebug v34.00.16

CMDebug is a stand-alone version of the Take Command IDE and batch debugger. CMDebug is intended to help developers create a batch file to run in CMD (the Windows default command processor) or TCC-RT (the free runtime version of TCC). It provides a comprehensive set of batch debugging features specifically designed for batch file development.

[Download JP Software CMDebug v34.00.16](https://developer.team/misc-development/35198-jp-software-cmdebug-v340016.html)